move types.ts
